+++ MeuDevocional/Dependences/ConstraintsAndDesign.swift
@@ -15,38 +15,6 @@ public func editaTituloVerde(titulo: UILabel){
     titulo.textColor = verde
 }
 
-///configurando a textField para receber apenas uma quantidade máxima de letras
-private var kAssociationKeyMaxLength: Int = 0
-extension UITextField {
-        @IBInspectable var maxLength: Int {
-            get {
-                if let length = objc_getAssociatedObject(self, &kAssociationKeyMaxLength) as? Int {
-                    return length
-                } else {
-                    return Int.max
-                }
-            }
-            set {
-                objc_setAssociatedObject(self, &kAssociationKeyMaxLength, newValue, .OBJC_ASSOCIATION_RETAIN)
-                addTarget(self, action: #selector(checkMaxLength), for: .editingChanged)
-            }
-        }
-
-        @objc func checkMaxLength(textField: UITextField) {
-            guard let prospectiveText = self.text,
-                prospectiveText.count > maxLength
-                else {
-                    return
-            }
-
-            let selection = selectedTextRange
-            let indexEndOfText = prospectiveText.index(prospectiveText.startIndex, offsetBy: maxLength)
-            let substring = prospectiveText[..<indexEndOfText]
-            text = String(substring)
-            selectedTextRange = selection
-        }
-}
-
 ///classe para validar os inputs do YouTube, Deezer, Spotify e Apple Music
 class Validation {
     public func validateYoutube(name: String) ->Bool {
