+++ MeuDevocional/Dependences/Notificacoes/NotificacoesConteudo.swift
@@ -1,0 +1,12 @@
+//
+//  Notifications.swift
+//  MeuDevocional
+//
+//  Created by Beatriz Duque on 05/01/22.
+//
+
+import Foundation
+
+///array com as notificações do app
+let notfTitles: [String] = ["Meu Devocional","Gratidão","Worship Time!","Salmo 37:5","Mateus 6:34","João 3:16","Colossenses 3:4","Momento de oração","Mateus 11:21","Mateus 22:37", "Romanos 8:38"]
+let notfContents: [String] = ["Não esqueça de anotar e compartilhar a devocional de hoje!","Já adicionou um motivo para ser grato no Mural hoje?","Que tal escolher uma devocional para leitura e louvor hoje?","'Entrega o teu caminho ao Senhor; confia nele, e ele tudo fará.'","...Portanto, não vos inquieteis com o dia de amanhã, pois o amanhã trará os seus cuidados; basta ao dia o seu próprio mal.","'Porque Deus tanto amou o mundo que deu o seu Filho Unigênito, para que todo o que nele crer não pereça, mas tenha a vida eterna.'","'Acima de tudo, porém, revistam-se do amor, que é o elo perfeito.'","Conta pra Ele, conte com Ele!","'Vinde a mim todos os que estais cansados e oprimidos, e eu vos aliviarei.'","'E Jesus disse-lhe: Amarás o Senhor, teu Deus, de todo o teu coração, e de toda a tua alma, e de todo o teu pensamento.'","'Nem a altura, nem a profundidade, nem alguma outra criatura nos poderá separar do amor de Deus, que está em Cristo Jesus, nosso Senhor!'"]
