+++ MeuDevocional/Dependences/UserDefaults+Devocionais/UserDevocionais.swift
@@ -1,0 +1,27 @@
+//
+//  UserDevocionais.swift
+//  MeuDevocional
+//
+//  Created by Beatriz Duque on 05/01/22.
+//
+
+import Foundation
+
+///base de dados que sera alterado pelo usuario
+let defaults = UserDefaults.standard
+var baseAplicacaoCotidiano1 = ["","","","",""]
+var baseAplicacaoCotidiano2 = ["","","","",""]
+var baseAplicacaoCotidiano3 = ["","","","",""]
+var baseAnotacaoCotidiano = ["","","","",""]
+
+
+var baseAplicacaoVida1 = ["","","","","",""]
+var baseAplicacaoVida2 = ["","","","","",""]
+var baseAplicacaoVida3 = ["","","","","",""]
+var baseAnotacaoVida = ["","","","","",""]
+
+
+var baseAplicacaoEstudos1 = ["","","","",""]
+var baseAplicacaoEstudos2 = ["","","","",""]
+var baseAplicacaoEstudos3 = ["","","","",""]
+var baseAnotacaoEstudos = ["","","","",""]
