+++ MeuDevocional/DevocionaisDiarias/DevocionalMusicViewController.swift
@@ -10,16 +10,13 @@ import WebKit
 
 class DevocionalMusicViewController: UIViewController,WKNavigationDelegate {
     var recebe = ""
-    
-    
     var webView: WKWebView!
     override func loadView() {
         webView = WKWebView()
         webView.navigationDelegate = self
         view = webView
-        //view.addSubview(webView)
-        //webView.topAnchor.constraint(equalTo: view.topAnchor).isActive = true
     }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         navigationController?.navigationBar.prefersLargeTitles = false
